Deal list
The current url (for the call to retrieve the deals) it’s www.google.es
and the json is hardcoded. When the server is up and running that will
change.
Deleted the min sdk version, we have to check that (to make it more
compatible).
Changed in the menu the string to “Entrar”, that should send to the
SignIn/SignUp activity.
The shop id is shown now, but that will change to ShopName, or to
nothing.